requirements.txt
Pillow
pylibdmtx
pywin32
